Move logger to helpers
